+++ packages/blocks-core/js/libs/wordpress/accordion-panel/test/functionality.blocks-1.e2e.cy.js
@@ -53,6 +53,15 @@ describe('Accordion Panel Block', () => {
 
 		cy.checkBlockCardItems(['normal', 'hover', 'core/paragraph']);
 
+		cy.checkBlockStatesPickerItems([
+			'elements/bold',
+			'elements/italic',
+			'elements/kbd',
+			'elements/code',
+			'elements/span',
+			'elements/mark',
+		]);
+
 		//
 		// 1. Edit Block
 		//